Reorganized and Standardized Naming
Reorganized components and cleaned up the naming conventions
