+++ frontend/components/admin/category/DeleteCategory.js
