+++ frontend/components/admin/meat/EditMeat.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Mutation } from 'react-apollo';
 import PropTypes from 'prop-types';
-import { Form } from './styles/Form';
-import { ErrorMessage } from './ErrorMessage';
-import { UPDATE_MEAT_MUTATION } from '../mutations/Meat';
-import { ALL_MEATS_QUERY } from '../queries/Meats';
+import { Form } from '../../styles/Form';
+import { ErrorMessage } from '../../ErrorMessage';
+import { UPDATE_MEAT_MUTATION } from '../../../mutations/Meat';
+import { ALL_MEATS_QUERY } from '../../../queries/Meat';
 
 class EditMeat extends Component {
     static propTypes = {
